pyproject.toml
[project]
name = "Petrichor"
version = "1.0.0"
description = "A silly bot for me and my friends."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "discord-py>=2.5.2",
]
